+++ ec/src/models/mod.rs
@@ -1,4 +1,4 @@
-use crate::ProjectiveCurve;
+use crate::{AffineCurve, ProjectiveCurve};
 use ark_ff::{Field, PrimeField, SquareRootField, Zero};
 
 pub mod bls12;
@@ -78,6 +78,15 @@ pub trait SWModelParameters: ModelParameters {
         Self::mul_affine(item, Self::ScalarField::characteristic()).is_zero()
     }
 
+    /// Performs cofactor clearing.
+    /// The default method is simply to multiply by the cofactor.
+    /// Some curves can implement a more efficient algorithm.
+    fn clear_cofactor(
+        item: &short_weierstrass_jacobian::GroupAffine<Self>,
+    ) -> short_weierstrass_jacobian::GroupAffine<Self> {
+        item.mul_by_cofactor()
+    }
+
     /// Default implementation of group multiplication for projective
     /// coordinates
     fn mul_projective(
@@ -146,6 +155,16 @@ pub trait TEModelParameters: ModelParameters {
         item: &twisted_edwards_extended::GroupAffine<Self>,
     ) -> bool {
         Self::mul_affine(item, Self::ScalarField::characteristic()).is_zero()
+    }
+
+    /// Performs cofactor clearing.
+    /// The default method is simply to multiply by the cofactor.
+    /// For some curve families though, it is sufficient to multiply
+    /// by a smaller scalar.
+    fn clear_cofactor(
+        item: &twisted_edwards_extended::GroupAffine<Self>,
+    ) -> twisted_edwards_extended::GroupAffine<Self> {
+        item.mul_by_cofactor()
     }
 
     /// Default implementation of group multiplication for projective